requirements.txt
fastapi
uvicorn
gradio
torch
imageio
numpy
Pillow
easydict
